# openapi.yaml
openapi: 3.1.0
info:
  title: Astradio API
  version: "1.0"
servers:
  - url: https://api.astradio.io/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email, password], properties: { email: {type: string}, password: {type: string, minLength: 12} } }
      responses: { "200": { description: OK } }
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [email, password], properties: { email: {type: string}, password: {type: string} } }
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema: { type: object, properties: { access_token: {type: string}, expires_in: {type: integer} } }
...